Refactor console.py and agent_fetcher.py
