+++ src/ostorlab/agent/agent_fetcher.py
@@ -8,11 +8,10 @@ import re
 import docker
 import docker.errors
 
-from ostorlab import configuration_manager
+from ostorlab import configuration_manager, console as cli_console
 from ostorlab.apis import agent_details as agent_details_api
 from ostorlab.apis.runners import public_runner, authenticated_runner
 from ostorlab.apis.runners import runner as base_runner
-from ostorlab.cli import console as cli_console
 from ostorlab.agent import definitions as agent_definitions
 from ostorlab.utils import version as version_definition
