+++ src/ostorlab/cli/agent/healthcheck/healthcheck.py
@@ -5,7 +5,7 @@ import logging
 import click
 import httpx
 
-from ostorlab.cli import console as cli_console
+from ostorlab import console as cli_console
 from ostorlab.cli.agent import agent
 
 console = cli_console.Console()
