+++ src/ostorlab/cli/agent/install/install.py
@@ -2,9 +2,9 @@
 
 import click
 
-from ostorlab.cli import agent_fetcher
+from ostorlab.agent import agent_fetcher
 from ostorlab.cli.agent import agent as agent_command
-from ostorlab.cli import console as cli_console
+from ostorlab import console as cli_console
 from ostorlab.cli import install_agent
 from ostorlab.cli import docker_requirements_checker
 from ostorlab.cli import types
