+++ src/ostorlab/cli/auth/revoke/revoke.py
@@ -5,9 +5,8 @@ import logging
 from ostorlab.apis.runners import authenticated_runner
 from ostorlab.apis.runners import runner
 from ostorlab.apis import revoke_api_key
-from ostorlab import configuration_manager
+from ostorlab import configuration_manager, console as cli_console
 from ostorlab.cli.auth import auth
-from ostorlab.cli import console as cli_console
 
 logger = logging.getLogger(__name__)
 console = cli_console.Console()
