+++ src/ostorlab/cli/scan/run/assets/android_aab.py
@@ -9,8 +9,7 @@ import click
 
 from ostorlab.assets import android_aab as android_aab_asset
 from ostorlab.cli.scan.run import run
-from ostorlab.cli import console as cli_console
-from ostorlab import exceptions
+from ostorlab import exceptions, console as cli_console
 
 console = cli_console.Console()
 logger = logging.getLogger(__name__)
