+++ src/ostorlab/cli/scan/run/assets/file.py
@@ -8,9 +8,8 @@ from typing import Optional, Tuple
 import click
 
 from ostorlab.assets import file as file_asset
-from ostorlab.cli import console as cli_console
 from ostorlab.cli.scan.run import run
-from ostorlab import exceptions
+from ostorlab import exceptions, console as cli_console
 
 console = cli_console.Console()
 logger = logging.getLogger(__name__)
