+++ src/ostorlab/cli/scan/run/assets/ios_ipa.py
@@ -10,8 +10,7 @@ import click
 
 from ostorlab.assets import ios_ipa as ios_ipa_asset
 from ostorlab.cli.scan.run import run
-from ostorlab.cli import console as cli_console
-from ostorlab import exceptions
+from ostorlab import exceptions, console as cli_console
 
 console = cli_console.Console()
 logger = logging.getLogger(__name__)
