+++ src/ostorlab/cli/vulnz/dump/dump.py
@@ -6,7 +6,7 @@ import click
 import sqlalchemy
 from ostorlab.apis.runners import runner
 
-from ostorlab.cli import console as cli_console
+from ostorlab import console as cli_console
 from ostorlab.cli import dumpers
 from ostorlab.cli.vulnz import vulnz
