+++ src/ostorlab/cli/vulnz/list/list.py
@@ -5,7 +5,7 @@ from typing import Optional
 
 import click
 
-from ostorlab.cli import console as cli_console
+from ostorlab import console as cli_console
 from ostorlab.cli.vulnz import vulnz
 
 console = cli_console.Console()
