+++ src/ostorlab/runtimes/local/runtime.py
@@ -20,10 +20,9 @@ from rich import markdown
 from rich import panel
 from sqlalchemy import case
 
-from ostorlab import exceptions
+from ostorlab import exceptions, console as cli_console
 from ostorlab.assets import asset as base_asset
-from ostorlab.cli import console as cli_console
-from ostorlab.cli import agent_fetcher
+from ostorlab.agent import agent_fetcher
 from ostorlab.cli import docker_requirements_checker
 from ostorlab.cli import dumpers
 from ostorlab.cli import install_agent
